fixed bug with current location btn
When current location button is clicked, the lat and lon of the current location display on the page, instead of the city name.
Reorganized the layout of the JS file to be more readable.
Made the F and C buttons functional (both showing active currently, to be fixed at a later date.)